+++ packages/dotnet/battlenet-lancache-prefill/default.nix
@@ -1,5 +1,7 @@
-{ lib, fetchFromGitHub, buildDotnetModule, dotnet-runtime }:
+{ lib, fetchFromGitHub, buildDotnetModule, dotnet-runtime_8, dotnet-sdk_8 }:
 let
+  dotnet-runtime = dotnet-runtime_8;
+  dotnet-sdk = dotnet-sdk_8;
   pname = "battlenet-lancache-prefill";
   version = "1.7.0";
 
@@ -23,11 +25,12 @@ let
 
   projectFile = [
     "BattleNetPrefill/BattleNetPrefill.csproj"
-    "LancachePrefill.Common/LancachePrefill.Common.csproj"
+    "LancachePrefill.Common/dotnet/LancachePrefill.Common.csproj"
   ];
 
   nugetDeps = ./deps.nix;
 
 in buildDotnetModule {
-  inherit pname version meta src dotnet-runtime patches projectFile nugetDeps;
+  inherit pname version meta src dotnet-runtime dotnet-sdk patches projectFile
+    nugetDeps;
 }